+++ packages/react/src/table/TableHeader.tsx
@@ -3,7 +3,6 @@ import { Box } from '../box';
 import { ResponsiveProp } from '../core';
 
 export type TableHeaderProps = PropsWithChildren<{
-	'aria-label'?: string;
 	/** Can be used to conditionally hide or show table cells at different breakpoints. */
 	display?: ResponsiveProp<'none' | 'table-cell'>;
 	/** Defines the cells that the header (defined in the <th>) element relates to. */
@@ -28,7 +27,9 @@ export const TableHeader = ({
 			fontWeight="bold"
 			focus
 			width={width}
-			css={{ textAlign }}
+			css={{
+				textAlign,
+			}}
 			{...props}
 		>
 			{children}
